Remove old command line files
