+++ clowder/clowder/cli/diff_controller.py
@@ -12,7 +12,7 @@ from clowder.clowder_repo import (
     valid_clowder_yaml_required
 )
 from clowder.cli.globals import CLOWDER_CONTROLLER
-from clowder.commands.util import (
+from clowder.cli.util import (
     filter_groups,
     filter_projects_on_project_names,
     run_group_command,
