+++ clowder/clowder/cli/prune_controller.py
@@ -17,7 +17,7 @@ from clowder.clowder_repo import (
     valid_clowder_yaml_required
 )
 from clowder.cli.globals import CLOWDER_CONTROLLER
-from clowder.commands.util import (
+from clowder.cli.util import (
     existing_branch_groups,
     existing_branch_projects,
     filter_groups,
