--- clowder/clowder/commands/__init__.py
@@ -1,23 +1,0 @@
-from clowder.commands.branch import branch
-from clowder.commands.checkout import checkout
-from clowder.commands.clean import (
-    clean,
-    clean_all
-)
-from clowder.commands.diff import diff
-from clowder.commands.parallel import (
-    forall,
-    herd,
-    herd_parallel,
-    reset,
-    sync
-)
-from clowder.commands.prune import prune
-from clowder.commands.save import save
-from clowder.commands.start import (
-    start_groups,
-    start_projects
-)
-from clowder.commands.stash import stash
-from clowder.commands.status import status
-from clowder.commands.yaml import yaml
