--- docs/clowder.commands.rst
@@ -1,110 +1,0 @@
-clowder\.commands package
-=========================
-
-Submodules
-----------
-
-clowder\.commands\.branch module
---------------------------------
-
-.. automodule:: clowder.commands.branch
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.checkout module
-----------------------------------
-
-.. automodule:: clowder.commands.checkout
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.clean module
--------------------------------
-
-.. automodule:: clowder.commands.clean
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.diff module
-------------------------------
-
-.. automodule:: clowder.commands.diff
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.parallel module
-----------------------------------
-
-.. automodule:: clowder.commands.parallel
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.prune module
--------------------------------
-
-.. automodule:: clowder.commands.prune
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.save module
-------------------------------
-
-.. automodule:: clowder.commands.save
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.start module
--------------------------------
-
-.. automodule:: clowder.commands.start
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.stash module
--------------------------------
-
-.. automodule:: clowder.commands.stash
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.status module
---------------------------------
-
-.. automodule:: clowder.commands.status
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.util module
-------------------------------
-
-.. automodule:: clowder.commands.util
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-clowder\.commands\.yaml module
-------------------------------
-
-.. automodule:: clowder.commands.yaml
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: clowder.commands
-    :members:
-    :undoc-members:
-    :show-inheritance:
