+++ script/clean
@@ -13,8 +13,8 @@ clean_clowder_repo() {
     sudo -H rm -rf $clowder_path/build
     echo "Removing .pyc and __pycache__"
     local path=$clowder_path/clowder
-    sudo -H rm $path/*.pyc $path/*/*.pyc
-    sudo -H rm -rf $path/__pycache__ $path/*/__pycache__
+    sudo -H rm $path/*.pyc $path/*/*.pyc $path/*/*/*.pyc
+    sudo -H rm -rf $path/__pycache__ $path/*/__pycache__ $path/*/*/__pycache__
     echo "pip uninstall clowder"
     pushd $REPO_PATH || exit 1
     sudo -H pip2 uninstall clowder-repo
